Scm definition moved to Error
